+++ package.json
@@ -6,36 +6,41 @@
     "dev": "vite dev",
     "build": "vite build",
     "preview": "vite preview",
-    "vercel-build": "prisma generate && yarn build"
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+    "vercel-build": "prisma generate && yarn build",
+    "lint": "prettier --plugin-search-dir . --check . && eslint ."
   },
   "devDependencies": {
-    "@sveltejs/adapter-vercel": "1.0.0-next.66",
-    "@sveltejs/kit": "1.0.0-next.405",
-    "@typescript-eslint/eslint-plugin": "^5.33.0",
-    "@typescript-eslint/parser": "^5.33.0",
-    "@unocss/preset-typography": "^0.45.6",
-    "@unocss/preset-wind": "^0.45.6",
-    "@unocss/reset": "^0.45.6",
-    "@unocss/transformer-variant-group": "^0.45.6",
-    "eslint": "^8.22.0",
+    "@sveltejs/adapter-vercel": "1.0.0-next.79",
+    "@sveltejs/kit": "1.0.0-next.516",
+    "@typescript-eslint/eslint-plugin": "^5.40.0",
+    "@typescript-eslint/parser": "^5.40.0",
+    "@unocss/preset-typography": "^0.45.29",
+    "@unocss/preset-wind": "^0.45.29",
+    "@unocss/reset": "^0.45.29",
+    "@unocss/transformer-variant-group": "^0.45.29",
+    "eslint": "^8.25.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-svelte3": "^4.0.0",
     "gray-matter": "^4.0.3",
     "mdsvex": "^0.10.6",
-    "prisma": "^4.2.1",
+    "prettier": "^2.7.1",
+    "prettier-plugin-svelte": "^2.8.0",
+    "prisma": "^4.4.0",
     "rehype-autolink-headings": "^6.1.1",
     "rehype-slug": "^5.0.1",
-    "svelte": "^3.49.0",
-    "svelte-check": "^2.8.0",
+    "svelte": "^3.52.0",
+    "svelte-check": "^2.9.2",
     "svelte-preprocess": "^4.10.7",
     "tslib": "^2.4.0",
-    "typescript": "~4.7.4",
-    "unocss": "^0.45.6",
-    "vite": "^3.0.7"
+    "typescript": "~4.8.4",
+    "unocss": "^0.45.29",
+    "vite": "^3.1.8"
   },
   "type": "module",
   "dependencies": {
-    "@fontsource/montserrat": "^4.5.12",
-    "@fontsource/open-sans": "^4.5.11",
-    "@prisma/client": "^4.2.1"
+    "@fontsource/montserrat": "^4.5.13",
+    "@fontsource/open-sans": "^4.5.13",
+    "@prisma/client": "^4.4.0"
   }
 }
