+++ src/lib/store/guest.ts
@@ -1,5 +1,5 @@
 import { writable } from 'svelte/store';
 
-import type { Guest } from 'src/routes/api/guests.json';
+import type { Guest } from 'src/routes/api/guests/+server';
 
 export const guestsStore = writable<Guest[]>([]);
