+++ src/routes/+page.svelte
@@ -1,24 +1,11 @@
-<script context="module" lang="ts">
-  export async function load({ fetch }) {
-    const getPosts = await fetch('/api/posts.json');
-    const posts = (await getPosts.json()) as Post[];
-
-    return {
-      props: {
-        posts
-      }
-    };
-  }
-</script>
-
 <script lang="ts">
   import Divider from '$lib/components/Divider.svelte';
   import PostList from '$lib/components/PostList.svelte';
   import Seo from '$lib/components/SEO.svelte';
-  import type { Post } from '$lib/utils/fetch-data';
   import logo from '../lib/assets/logo.jpg';
 
-  export let posts: Post[];
+  /** @type {import('./$types').PageData} */
+  export let data;
 </script>
 
 <Seo title="Home" />
@@ -31,10 +18,14 @@
   Terkadang memang suka menulis hal-hal yang random.
 </section>
 
-<a class="mt-8 mb-16 font-semibold text-blue-500 bg-gray-200 p-1 rounded-sm w-fit" href="/guestbook">👤 Buku tamu</a>
+<a
+  class="mt-8 mb-16 font-semibold text-blue-500 bg-gray-200 p-1 rounded-sm w-fit"
+  href="/guestbook"
+  data-sveltekit-prefetch>👤 Buku tamu</a
+>
 
 <Divider />
 
 <h2 id="tulisan" class="text-3xl font-bold mt-16 leading-relaxed">🌅 Tulisan</h2>
 
-<PostList {posts} />
+<PostList posts={data.posts} />
