--- src/routes/tags/[tag].svelte
@@ -1,32 +1,0 @@
-<script context="module" lang="ts">
-  export async function load({ fetch, params }) {
-    const getPosts = await fetch('/api/posts.json');
-    const posts = (await getPosts.json()) as Post[];
-    const tag = params.tag as string;
-
-    const postsByTag = posts.filter(post => post.tags.includes(deslugify(tag)));
-
-    return {
-      props: {
-        posts: postsByTag ?? [],
-        tag: deslugify(tag)
-      }
-    };
-  }
-</script>
-
-<script lang="ts">
-  import PostList from '$lib/components/PostList.svelte';
-  import Seo from '$lib/components/SEO.svelte';
-  import type { Post } from '$lib/utils/fetch-data';
-  import { deslugify } from '$lib/utils/slug';
-
-  export let posts: Post[];
-  export let tag: string;
-</script>
-
-<Seo title={tag} />
-
-<h1 class="text-4xl font-bold leading-relaxed">🧵{tag}</h1>
-
-<PostList {posts} />
